Output has predictions and stats in dictionary
