+++ esports/api/fast_api.py
@@ -5,11 +5,16 @@ from xgboost import XGBClassifier
 from sklearn.preprocessing import RobustScaler
 import pandas as pd
 
+# ignore warnings
+import warnings
+warnings.simplefilter(action='ignore', category=FutureWarning)
+pd.options.mode.chained_assignment = None
+
 # local imports
 from esports.scrape.steam_id_finder import steam_id_finder
-from esports.transform_data import average_player_data
-from esports.get_wl_data import get_wl_data
 from esports.matches_clean import clean_player_pairs_data, train_test_split_data
+from esports.preprocess import preprocess_pairs
+from esports.model.XGB_model import xgb_model
 
 app = FastAPI()
 
@@ -42,36 +47,8 @@ def predict(account_id, opponent_id):
 
     # ~~~~~~ Win rate data~~~~~~~~
     # make sure to retrieve the win rate data here to pass onto the machine learning model
-    player_avg = average_player_data(user_steam_id)
-    opponent_avg = average_player_data(opps_steam_id)
-    player_wl = get_wl_data(user_steam_id)
-    opponent_wl = get_wl_data(opps_steam_id)
 
-    keys_to_transform = [
-        'kills_per_min', 'deaths_per_min', 'assists_per_min', 'xp_per_min',
-        'gold_per_min', 'hero_damage_per_min', 'tower_damage_per_min',
-        'last_hits_per_min'
-    ]
-
-    for key in keys_to_transform:
-        player_avg = player_avg.rename(index={key: f'player_{key}'})
-        opponent_avg = opponent_avg.rename(index={key: f'opponent_{key}'})
-
-    # Append win_ratio to player_avg and opponent_avg
-    player_win_ratio = pd.Series(
-        [player_wl["win"] / (player_wl["win"] + player_wl["lose"])],
-        index=['player_win_ratio'])
-
-    opponent_win_ratio = pd.Series(
-        [opponent_wl["win"] / (opponent_wl["win"] + opponent_wl["lose"])],
-        index=['opponent_win_ratio'])
-
-    X_pred = player_avg.append(opponent_avg)
-    X_pred = X_pred.append(player_win_ratio)
-    X_pred = X_pred.append(opponent_win_ratio)
-
-    X_pred = pd.DataFrame(X_pred)
-    X_pred = X_pred.T
+    X_pred = preprocess_pairs(user_steam_id, opps_steam_id)
 
     # ~~~~~~~~~~~ RUN MODEL ~~~~~~~~~~~~~~~
     # replace with actual model
@@ -79,36 +56,14 @@ def predict(account_id, opponent_id):
     #    return 0.5
     #prediction = model()
 
-    def xgb_model(X_pred):
-        df = clean_player_pairs_data()
-        X = df.drop(columns=[
-            'match_id', 'player', 'opponent', 'winner',
-            'player_roshans_killed', 'player_obs_placed',
-            'opponent_roshans_killed', 'opponent_obs_placed', 'player_win',
-            'player_tower_damage', 'opponent_tower_damage'
-        ])
-
-        y = df['player_win']
-
-        # train test split
-        X_train, X_test, y_train, y_test = train_test_split_data()
-
-        # Scale numerical variables
-        rb_scaler = RobustScaler()
-        X_train_scaled = rb_scaler.fit_transform(X_train)
-        X_test_scaled = rb_scaler.transform(X_test)
-
-        xgb = XGBClassifier()
-
-        xgb.fit(X_train_scaled, y_train)
-        X_pred = rb_scaler.transform(X_pred)
-        y_pred = xgb.predict_proba(X_pred)
-
-        return y_pred
-
     prediction = xgb_model(X_pred)
+    #print(type(X_pred))
 
-    return (prediction)
+    output = {'player_pred':prediction[0][1],
+              'opponent_pred':prediction[0][0],
+              'stats':X_pred.to_dict(orient='records')}
+
+    return output
 
 
 print(predict(148673797, 392047872))
